+++ utils/index.ts
@@ -1,2 +1,1 @@
 export * from './axiosInstance';
-export * from './useAsync';
